Fix with_confirmation on Windows
File ids where different when communication back within a confirmation
so that they could not be matched anymore. This happened when the
receiver was running on Linux and the sender on Windows. This was fixed
by converting communicated paths to posix format (like on Linux) by
using pathlib independently where the sender was running on.