Add workaround for log file rotation on Windows
Rotating the log files does not work on Windows (reason unknown) but a
workaround was needed to prevent the files from growing too big. Thus
the possibility was added to add the date (as placeholder) to the
log name, resulting in a new log file on restart of the datamanager on
another day.