Fix unittests for signalhandler and datadispatcher
After the last hotfixes unittests were broken. The datadispatcher test
did not stop anymore due to the added datahandling thread in the
datadispatcher which had to be stopped over the control socket.
Additionally the signalhandler tests had to be modified due to the
migration of the signal checks into the base class.