Fix hidra usage via docker
The dockerfiles have to be updated for the new dependencies. In addition
to be able to modify the config file used inside the docker container
hidra is not cloned inside the container anymore but rather mapped into
it. Due to the special setup of docker inside the desy environment
additional parameters are needed and to handle this convinently a
hidra_docker script is added.