Fix task sending if datadispatcher are busy
If the datadispatchers are busy with processing data and thus the timeout
is reached in the task sending socket (router socket), this would lead
to a shutdown of hidra. As remedy the control socket is checked now
after a such a timeout. If everything is fine there, the message sending
to the router socket is retried.