Fix hanging processes due of missed control signal
If one of the first started processes stops (e.g. event detector because
of wrong config), later started datadispatcher hung (especially if a
bigger number worker was configured). The cause lies in the zmq slow
joiner problem occuring because the exit signals are sent before the
later processes are started.