Fix stopping running instances
Previously, the stop method was always called on a new controller instance, even
if one already existed, leaving the already existing instance in an inconsistent
state.

In particular, the existing instance did not refresh its config cache when Hidra
was restarted, resulting in wrong port numbers being served to clients.