Build freeze packages in manylinux1 container
By using the schooft/manylinux_freeze:2.7 docker image that was derived from the
manylinux1 project and includes a shared build of Python, cx_Freeze should in
principle be able to create packages that work on any platform supported by the
manylinux1 project.

Because the manylinux1 docker image does not run on modern Debian systems, a
vagrant file is provided to start a Centos 7 VM that supports running the
manylinux1 docker image.

Some changes to freeze_setup.py are necessary to enable a successful build in
the container. Additionally, multiple event detectors are supported now as well.

A few dependency issues remain that prevent the package from running on all
manylinux1 compatible systems. The patchelf program can be used to fix such
cases when they affect targeted systems.