Add message index offset from regex capture group
Sometimes it is not possible to have a unique message index in the file
name. For these cases, a regex capture group named file_idx_offset can
be used. If the group is present in the regex, its value will be added
to the value of the file_idx_in_scan capture group to obtain the actual
message id for the asapo stream.