Limit number of open asapo producer connections
Each producer keeps an open connection for each producer thread. If a
lot unique data sources are created by a beamline, Hidra can exhaust the
maximum number of open files, making it impossible to save new files.

With this commit, the producer for the least recently used data source
is deleted when a maximum number of producers is exceeded. Currently,
the maximum number is hard-coded to 10 producers.